feat: add getRenderHandler & optimize types
